QEMU: support USB cdrom devices
If a USB cdrom is configured, the media type is silently ignored,
when generating the QEMU command, so libvirt actually generates
USB disks. The main problem is, that -blockdev mechanism relies
on the device type to handle the media type, like ide-cd and
ide-hd. But there is just usb-storage.

As a result the Windows 10 Arm64 installer won't find the virtio
ISO or even its own install media. You can actually turn them into
USB sticks, by setting them removable, so they show up in the
installer, but then it doesn't expect joliet or UDF filesystems,
so can't access these either.

So this generates the old driver+device commandline arguments in
the case of USB cdrom devices to make the installer happy.

Signed-off-by: Jan-Marek Glogowski <glogow@fbihome.de>
Message-Id: <20200906143615.30816-2-glogow@fbihome.de>