meson: Add tests_timeout build option
The default timeout for tests is 30s, but that's not always
enough time: the Valgrind test, for example, are currently
special-cased because they take longer, and on certain
architectures where powerful hardware is not (yet) available
even the basic test programs might take more than 30s to
complete.

A concrete example of this happening and preventing a package
build from finishing:

  https://buildd.debian.org/status/fetch.php?pkg=libvirt&arch=riscv64&ver=6.7.0-1&stamp=1599587548&raw=0

Make it possible to configure a different timeout for tests.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Message-Id: <20200909131006.128717-3-abologna@redhat.com>