tests: don't mix FILE* and UNIX FD I/O on same stream
There is currently a hand in test27 that exhibits itself on FreeBSD 11.4
only. The behaviour is that virCommandProcessIO gets POLLIN on the
FD for stdout, but read() blocks. Meanwhile commandtest also blocks
in write for stderr because the pipe buffers are full.

This fix in commandhelper likely does not really address the root cause
just hides it due to the buffering done by FILE *. Mixing UNIX FD I/O
and FILE * I/O is bad practice regardles.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>
Message-Id: <20200921173653.1986867-1-berrange@redhat.com>
Reviewed-by: Eric Blake <eblake@redhat.com>