qemu: fix error message when baselining with a single cpu
When executing the hypervisor-cpu-baseline command and if there is
only a single CPU definition present in the XML file, then libvirt
will print an unhelpful message:

"error: An error occurred, but the cause is unknown"

This is due to no CPU definition ever being "baselined", since the
API expects at least two CPU models.

Let's fix this by performing a CPU model expansion on the single CPU
definition and returning the result to the caller. This will also
ensure the CPU model is sane.

Signed-off-by: Collin Walling <walling@linux.ibm.com>
Message-Id: <20200925002239.61628-6-walling@linux.ibm.com>