lib: Don't unref message passed to virGDBusCallMethod{WithFD}()
The virGDBusCallMethod() and virGDBusCallMethodWithFD() are
simple wrappers over g_dbus_connection_call_sync() and
g_dbus_connection_call_with_unix_fd_list_sync() respectively. The
documentation to these function states that passed parameters
(@message in our case) is consumed for 'convenient' inline use of
g_variant_new() [1]. But that means we must not unref the message
afterwards. To make it explicit that the message is consumed the
signature of our wrappers is changed too.

1: https://developer.gnome.org/gio/stable/GDBusConnection.html#g-dbus-connection-call-sync

Reported-by: Cole Robinson <crobinso@redhat.com>
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <9ab312eb34cf10641de9b097d7a656e8fc2fce7f.1601630385.git.mprivozn@redhat.com>