qemu: Don't generate '-machine memory-backend' and '-numa memdev'
In 88957116c9 I've switched to -machine memory-backend=ID and
-object memory-backend-* because QEMU is obsoleting -mem-path
and -mem-prealloc. However, what I did not foresee was that using
-machine memory-backend in combination with -numa is not allowed
in QEMU. This was reported upstream and fortunately not released
yet.

The solution is to generate -machine memory-backend=ID iff there
are no guest NUMA nodes, and require use of -numa memdev= for new
enough QEMUs.

Reported-by: Masayoshi Mizuma <msys.mizuma@gmail.com>
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <5b20fc193a040a44009bb52ee87f462c3e62d270.1601971409.git.mprivozn@redhat.com>