tests: Rename some test files in cputestdata
The files contained the "-invalid" marker in their filename, marking
them as test cases that are supposed to fail. Unfortunately, the
"-invalid" marker does not discriminate between different tests the
files might be used in.

A later patch will introduce a new test validating the XML. This
test is not supposed to fail, as the files contain valid XML.

Signed-off-by: Tim Wiederhake <twiederh@redhat.com>
Message-Id: <20201007085458.231179-4-twiederh@redhat.com>